services/processing_service/requirements.txt
kafka-python
psycopg2-binary
influxdb-client
